Replace Shindo integration tests with Minitest
