+++ fog-dnsimple.gemspec
@@ -21,7 +21,6 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency "bundler", "~> 1.12"
   spec.add_development_dependency "rake", "~> 12.0"
-  spec.add_development_dependency "shindo", "~> 0.3"
   spec.add_development_dependency "minitest"
   spec.add_development_dependency "minitest-reporters"
