+++ tests/dns/models/record_test.rb
@@ -1,4 +1,4 @@
-require 'test_header'
+require 'test_helper'
 
 class Fog::DNS::Dnsimple::RecordTest < Minitest::Test
   include TestModelHelpers
