+++ tests/helpers/fog_helper.rb
@@ -1,15 +1,3 @@
-require "minitest/autorun"
-require "minitest/reporters"
-
-Minitest::Reporters.use! Minitest::Reporters::DefaultReporter.new(color: true)
-
-$LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
-require "fog/dnsimple"
-
-ENV["FOG_MOCK"] ||= "true"
-
-require_relative "helpers/mock_helper"
-
 module TestModelHelpers
 
   # Generates a unique identifier with a random differentiator.
